Rework when writer is opened
side effect here is that same file is set back to itself as tests are pretty dependent on somewhat bad design.  The code generally is designed well but tests are the bad aspect where they are too reliant on source.  The underlying getter in the writer really should be deleted, its final and can only be set by this class so there is zero reason it has a getter just so tests can look at content.